Added Utilities::powerspectrum2d() with float entries.

+++ include/utilities_slsim.h
@@ -1326,8 +1326,8 @@ void powerspectrum2dprebin(
 void powerspectrum2d(
                      std::valarray<float> const &aa      /// first realspace map to be
                      ,std::valarray<float> const &bb     /// second realspace map, same as aa to get power spectrum
-                     ,int nx                       /// number of pixels in x direction
-                     ,int ny                       /// number of pixels in y direction
+                     ,long nx                       /// number of pixels in x direction
+                     ,long ny                       /// number of pixels in y direction
                      ,double boxlx                 /// range of image in x direction
                      ,double boxly                 /// range of image in y direction
                      ,std::vector<double> &ll      /// output multiplot number of bins
@@ -1337,8 +1337,8 @@ void powerspectrum2d(
 
 void powerspectrum2d(
                      std::valarray<float> &aa      /// first realspace map to be
-                     ,int nx                       /// number of pixels in x direction
-                     ,int ny                       /// number of pixels in y direction
+                     ,long nx                       /// number of pixels in x direction
+                     ,long ny                       /// number of pixels in y direction
                      ,double boxlx                 /// range of image in x direction
                      ,double boxly                 /// range of image in y direction
                      ,std::vector<double> &ll      /// output multiplot number of bins
